main/requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
openai>=1.0.0
python-dotenv>=0.19.0
ultralytics>=8.0.0
streamlit>=1.18.0
pillow>=9.0.0 